Удалил лишний класс TryFigure (непонятное название - непонятная функция) Перенёс часть вспомогательных методов в статический класс Converter
