+++ AreaFigure.Common/Helpers/Converter.cs
@@ -1,0 +1,36 @@
+﻿using AreaFigure.Common.Figures;
+using System;
+using System.Linq;
+
+namespace AreaFigure.Common.Helpers
+{
+    public static class Converter
+    {
+        public static double[] ToDoubleArray(string str)
+        {
+            if (str == "")
+            {
+                return new double[0];
+            }
+
+            str = str.Trim();
+            var value = TrySplitToArray(str).Select(double.Parse).ToArray();
+
+            return Figure.CollapseNullValues(value);
+        }
+
+        public static string[] TrySplitToArray(string str)
+        {
+            try
+            {
+                var value = str.Split();
+                return value;
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+                throw;
+            }
+        }
+    }
+}
